Misc stuff
Yeah I know, the Gnollhack devs couldn't come up with a better commit title either. But anyway, who digs through commits looking for specific stuff? No one. And on the off chance that someone is still reading this, well, just look at the contents of the commit to see what it does :-P