Failsafe for off-hand lightsabers
Weird, couldn't reproduce the bug locally; what happened was that in my game on the server, I was dual-wielding and both the sabers were on, then I threw the primary one at a monster and it didn't come back but despite me no longer dual-wielding as a result, the secondary lightsaber didn't turn itself off.

Let's hope that this catchall works as intended.