Traps don't replace stairs
Yeah amateurhour, I guess you won after all... because it can happen that pits and the like are made as the game goes on, and they can also appear on just about any staircase, not just in Sokoban. If it was only Sokoban it'd be kinda okay, but as it is, it screws with the player too much and therefore has to get fixed. Argh.