Fix ceiling collapse trap
Sporkhack bug, discovered by shadowrider38. Oddly enough, it doesn't happen on all levels that allow the player to walk up to the map edge; it might depend on what's in memory when the game checks the "invalid" squares. If the memory contains code that corresponds to some terrain other than ROOM or CORR, it won't try to drop a rock on the invalid location and therefore probably won't crash.