Minimum value for stats is 1 now
This has been planned for a while. After all, why should they have a minimum of 3? That makes no sense. You get quite the debuffs if they are below 2, but on the bright side this change also works as a buffer against stat instadeath (currently only INT and WIS).

Starting stats for all player characters are still always at least 3. If we made them start at 1 sometimes, the buffer would once again not exist if you started with an INT of 1!