Fix amulet dropping exploits
Sorry shadowrider38, but it's far too easy to get rid of the fake amulet that way, especially since there's stuff like the ring of disarming which can just be swapped in to instantly drop the amulet whenever you want. That's obviously not intended, so this loophole is being closed now.