Thwart some recursion cheating
Of course no one has discovered the possibility of starting as, I dunno, tumblrer mahou shoujo, startscumming for the spellbook of character recursion and becoming something that doesn't have the tumblrer downside while keeping all the benefits. But just the mere thought that it could be exploited is unbearable for me, so here's a fix. Players have exploited every loophole they found in the past, after all, so who am I to assume they'd stop at this one? Hell no, if they figure it out they'll break the game again! ("break the game" = "render the game so easy that they can ascend it in their sleep without putting in any effort") ("ascend it in their sleep" = "ascend it more easily than I intended it to be")