Load others' savegames
This is annoying as heck, sometimes I have to transfer savegames for debug purposes. And I fucking HATE HATE HATE the programmers who made it so that the savegame is just erased unconditionally. Anything that does so shall die in a fire! Way too many functions like this one are scattered around the source code, the most obvious offender ("trickery", yeah right, as if that ever fired when you really cheated) was fixed long ago but this one remained until this day...