Finally add a way to recall
My guess is that vanilla nethack worked without it simply because it's a relatively short game, at least in absolute terms, and the devs therefore could get away with the stupid confused tele scrolls (which fail every other time and cannot go past the valley, let alone into branches), but this is SLEX. A faster and more reliable method of travelling is needed, badly. Especially since there's gonna be more branches added to the game eventually.