Tone down some bosses
Fuck you if you dislike this change! This is my game and I change whatever I want to :P The deadly sins were just too deadly in slex because in the original patch you weren't supposed to encounter all seven of them at once, so they need to be balanced in a way that amkes them not OP if they do spawn all at once :P And the uberversions of those alignment quest bosses should also be less deadly so that you don't get turboowned just because you didn't find the stupid portals.