Weather effects framework
The actual weather effects aren't coded yet but will be soon. Also I know it doesn't make sense for weather to exist in a dungeon but this is SLEX, it doesn't have to make sense! :P Maybe the ceiling has become porous or something, allowing the rain to pour through!