Address super user-unfriendly interface
Like what the fuck man, no one can figure that one out on their own! Players seem to finally not get stuck on their inability to turn the number pad on anymore, but now they get stuck on the look command, not realizing that of course they need to press comma, not space, when at the target location! Of course it's second nature for me and I'd generally not dick around trying to use the space bar, but that's because I know what the commands are while a new player usually doesn't...