New artifacts
Guh, that took a shitload of time. And I still don't have all the base items of the spellbook class covered! :( Many of this batch of artifacts were suggested by ChatGPT because I was having trouble coming up with enough ideas of my own; they may be cool and all but a lot of the bot's suggested artifacts were an unholy pain in the butt to actually implement!