Fix techs for metronome effect
Currently the tech metronome effect has only one source (a specific artifact that is very rare) but that is supposed to change. Either way, if it rolls an artifact that requires specific equipment, it should be possible for the player to quickly equip the required type of equipment so that the tech actually works.