Invoke deity change
It's really silly if it does rnz damage upon a failure. Since you usually use the tech when you need HP, a failure may well kill you anyway so those failures should at least not do excessive damage! Give the player a chance of survival!

Yeah, I originally came up with the tech, but that was back in the day when my sense of game balance was much worse than it is now so please forgive me ;) *angelic smile*