Segfault with item-stealing monsters
This doesn't really fix it, because the stealing monster still gets a bugged glorkum as a weapon. But at least the bugged glorkum now gets nulled out before something can try to do faulty calculations on it, like checking the small damage of the item with base type -16785031.

Honestly, I blame SLASH'EM's weird MON_WEP system with the temporary "mw" field, fairly certain vanilla (and, yes, probably 3.6 and onward too) solve this in a much more elegant way. But even if they do, backporting that to what is essentially 3.31 code would be hell...