Thwart some more wish scumming
Because if this isn't done, we might as well make players start with a scroll of wishing :-P If you want to play SLEX, you have to play it properly, not scum your way to an easy victory! Especially if you want to ascend a sheep prostitute. Either you have what it takes to win such a hard char or you don't deserve to win.